Split featurizers inside contrib, move test_lstm to proper subdirectory
