+++ tests/metal/modules/test_lstm.py
